refactor(arkworks): move ops tests to crate tests
