+++ crates/arkworks/src/cp6_782.rs
@@ -1,8 +1,8 @@
-use super::{all_curve_three_operations_test, CurveBasicOperations};
+use super::CurveBasicOperations;
 use ark_cp6_782::{Fq6, Fr, G1Projective, G2Projective, CP6_782};
 use ark_ec::PairingEngine;
 use ark_ff::{test_rng, Field, One, PrimeField};
-use ark_std::ops::MulAssign;
+use ark_std::{ops::MulAssign, vec::Vec};
 use rand::Rng;
 use rustc_hex::FromHex;
 
@@ -10,11 +10,6 @@ impl CurveBasicOperations for CP6_782 {
     const G1_LEN: usize = 209;
     const G2_LEN: usize = 625;
     const SCALAR_LEN: usize = 48;
-}
-
-#[test]
-fn test_cp6_782() {
-    all_curve_three_operations_test::<CP6_782>();
 }
 
 #[test]
