+++ tests/Cargo.toml
@@ -7,10 +7,24 @@ edition = "2018"
 [dev-dependencies]
 matter_labs = { package = "megaclite-matter-labs", path = "../crates/matter-labs" }
 zcash = { package = "megaclite-zcash", path = "../crates/zcash" }
+arkworks = { package = "megaclite-arkworks", path = "../crates/arkworks" }
+
+# Git deps
+ark-bls12-377 = { git = "https://github.com/arkworks-rs/curves", default-features = false, features = ["curve"] }
+ark-bls12-381 = { git = "https://github.com/arkworks-rs/curves", default-features = false, features = ["curve"] }
+ark-bn254 = { git = "https://github.com/arkworks-rs/curves", default-features = false, features = ["curve"] }
+ark-bw6-761 = { git = "https://github.com/arkworks-rs/curves", default-features = false }
+ark-cp6-782 = { git = "https://github.com/arkworks-rs/curves", default-features = false }
+ark-ff = { git = "https://github.com/arkworks-rs/algebra", default-features = false }
+ark-ec = { git = "https://github.com/arkworks-rs/algebra", default-features = false }
+ark-std = { git = "https://github.com/arkworks-rs/utils", default-features = false }
+ark-serialize = { git = "https://github.com/arkworks-rs/algebra", default-features = false }
+sapling_crypto_ce = { package = "sapling-crypto_ce", git = "https://github.com/matter-labs/sapling-crypto", branch = "master" }
+
+# crates.io deps
+rustc-hex = { version = "2", default-features = false }
 bellman_ce = { version = "0.3.0", default-features = false }
 num-bigint = "0.3"
 num-traits = "0.2"
 rand = "0.4"
 rust-crypto = "0.2.36"
-sapling_crypto_ce = { package = "sapling-crypto_ce", git = "https://github.com/matter-labs/sapling-crypto", branch = "master" }
-rustc-hex = { version = "2", default-features = false }
