+++ tests/src/matter_labs/mod.rs
@@ -1,3 +1,4 @@
+//! Matter Labs Tests
 pub mod altbn_128;
 pub mod parse;
 pub mod verify;
